Set default Astro `site`
The `site` attribute is important for the canonical URL and sitemap.
This commit will set the default `site` for deployment with Vercel.
For GitHub Pages deployments to https://br3ndonland.github.io/, the
GitHub Actions workflow uses `--site "https://${GITHUB_REPOSITORY##*/}"`
to override the URL in the Astro config file.

https://docs.astro.build/en/guides/configuring-astro/
https://docs.astro.build/en/reference/configuration-reference/