Filter socials in nav
The navigation menu contains social links including CV and resume.
CV and resume are currently using the same icon, which can be confusing.
The row of icons sometimes wraps on small screens like iPhone mini.
This commit will filter the CV link out of the nav. The CV link will
still be available in the footer.