Remove background noise
The CSS for light mode includes some background noise. An image is used
to produce repeated gray blocks, resulting in a textured appearance.
This background noise is difficult to control and evaluate. For example,
the CSS `background-blend-mode` property seems to be applied differently
in browsers on iOS than in browsers on macOS. The background noise seems
to be rendered as normal instead of overlay. To avoid this unnecessary
complexity, this commit will remove the background noise CSS.