Increase background gradient size
The background gradient at the top of the page has curves in it. These
curves were running across the nav text on some smaller mobile viewports
(around 375px), making the text difficult to see in light mode.

This commit will increase the background gradient size so that the curve
doesn't run across the text.