Prefetch pages in nav
This commit will add prefetch attributes to the internal site links in
the nav so that page navigation is quicker.

Nav URLs need to be adjusted for different trailing slash behavior on
GitHub Pages and Vercel. GitHub Pages uses trailing slashes and does not
allow configuration of this behavior, so attempting to prefetch URLs in
`consts.ts` that do not have trailing slashes will result in errors.
Vercel does allow configuration of trailing slash behavior with the
`cleanUrls` and `trailingSlash` configuration options, and these options
are set to disable trailing slashes in this project.

https://docs.astro.build/en/guides/prefetch/
https://vercel.com/docs/projects/project-configuration