Default to dark theme
This website includes both dark and light themes. When the site loads,
it defaults to the user's preferred theme, with a toggle to change
between dark and light.

This commit will update the theme toggle in `MainHead.astro` to default
to dark mode instead of to user preference. I prefer the colors in dark
mode and still need to work on some aspects of the color in light mode
(for example, black text on purple in the nav header).