Consolidate `about.astro` CSS grid in media query
The `/about` page uses CSS grid on larger viewports to move H2 headings
and their taglines out to the left margin of the page. In the `<style>`
section of `about.astro`, some of the grid styles were applied at the
top level, while others were applied with a media query. This resulted
in some styles being applied on smaller viewports when they didn't need
to be. The gap on the `.about` class was unnecessary on small viewports.

This commit will move the CSS grid styles into the media query so they
are only applied on larger viewports.