Always underline footer links on small viewports
The footer contains some anchor links. The anchor links have the same
color as the rest of the text. An underline appears on hover or focus.
On touch devices, hover or focus may not occur, and therefore the links
may be difficult to see.

This commit will update the footer to always underline anchor links on
smaller viewports. Behavior on larger viewports will stay the same.