Add Astro integration to autolink HTML headings
Commit aeb2ab8 added a Rehype plugin to automatically add anchor links
to headings from Markdown and MDX files. This heading autolinking works
well for Markdown, but does not process headings in non-Markdown files.
This commit will implement heading autolinking for non-Markdown files.

To accomplish this, an Astro integration will be added. This integration
will use the `astro:build:done` lifecycle hook to process files after
they have been built into HTML. The integration will accept a list of
file paths to transform. For each file, the integration will process the
HTML using the same Rehype plugin and configuration used for Markdown.
The result will be that heading autolinking will be performed on
Markdown and non-Markdown files in the same consistent way.

The integration currently only operates on the `astro:build:done` hook.
This means that heading autolinking will not be applied when starting
the Vite development server. In the future, additional build hooks could
be used to apply heading autolinking when running in development mode.

https://docs.astro.build/en/reference/integrations-reference/
https://github.com/rehypejs/rehype-autolink-headings
https://vitejs.dev/guide/api-plugin.html
https://rollupjs.org/plugin-development/