Move nav menu button to separate component
The navigation menu at the top of the page contains a menu button that
is visible on small viewports. The code for the button was mixed in with
the code for the rest of the nav. This commit will move the button to a
separate component without changing any of the functionality or styling.

Note that the button is still dependent on a separate HTML element with
id `menu-content` (in order to determine what content to toggle).