Move menu button class toggle back to nav
Commit 8669da6 moved the navigation menu button to a new component
without changing the code. The code for the menu button is more cleanly
separated from the rest of the nav, but further code changes are needed.

The web component custom element constructor for the button includes
logic to toggle a `menu-content` class on the navigation menu element.
The CSS class positions the menu on top of the rest of the page when the
menu is expanded on small viewports. Otherwise, the menu pushes the rest
of the page downward when expanded on small viewports. The CSS class is
added with JavaScript because different behavior is needed if JavaScript
is disabled. If JavaScript is disabled, the menu should be permanently
expanded and should not cover any other content on the page.

Neither the navigation menu nor the CSS for the `menu-content` class are
included in the code for the button itself, so the class toggle does not
need to be included in the button component code either.

This commit will move the class toggle from the button component back to
the nav component for better encapsulation of the logic for the button.

Note that the button is still dependent on a separate HTML element with
id `menu-content` (not class `menu-content`) in order to determine which
elements to toggle. This seems to be an acceptable practice, as this is
similar to what the Astro Starlight docs framework does. The Starlight
component `MobileMenuToggle.astro` looks for a parent `nav` element
outside the component (`this.closest("nav")`).

https://github.com/withastro/starlight/blob/bf939947a3b219a75cd7ac8ade27694c8a927937/packages/starlight/components/MobileMenuToggle.astro