Add menu button component prop for menu element ID
Commit 8669da6 moved the navigation menu button to a new component
without changing the code. The code for the menu button is more cleanly
separated from the rest of the nav, but further code changes are needed.

The menu button toggles visibility of the menu contents, but the button
component itself does not contain the code for the menu contents. The
button therefore needs to reference the separate HTML DOM element with
the menu contents in order to determine what to toggle. This seems to be
an acceptable practice, as this is similar to what the Astro Starlight
docs framework does. The Starlight component `MobileMenuToggle.astro`
looks for a parent `nav` element (`this.closest("nav")`). On this site,
the menu contents are not in a parent element, so a different selector
can be used, such as the element ID. The ID was previously hard-coded as
`menu-content`. It would be preferable to allow customization of the ID.

This commit will enable customization of the HTML element ID containing
the menu contents. To do so, a prop will be added to the menu button
component, with the previous value `"menu-content"` as default. A data
attribute (`data-*`) will be used to pass the ID in to the HTML custom
element for the menu button. The data attribute is then referenced in
the JavaScript class constructor for the menu button element through the
dataset property. Note the syntax considerations when working with data
attributes and the dataset property. In HTML, the data attribute name
begins with `data-` and is automatically casefolded. In JavaScript, the
`data-` prefix is automatically removed and the rest of the name is
converted from `lowercase-with-dashes` to `camelCase`. Therefore, the
data attribute added here is named `data-menu-element-id` in HTML but
referenced as `menuElementId` in JavaScript.

https://developer.mozilla.org/en-US/docs/Learn/HTML/Howto/Use_data_attributes
https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/dataset
https://developer.mozilla.org/en-US/docs/Web/API/Web_components/Using_custom_elements
https://docs.astro.build/en/guides/client-side-scripts/