Remove unnecessary CSS class from menu button
Move menu button class toggle back to nav

Commit 8669da6 moved the navigation menu button to a new component
without changing the code. The code for the menu button is more cleanly
separated from the rest of the nav, but further code changes are needed.

Now that the button is in its own component, there is no need to add a
separate CSS class `"menu-button"` to style the `<button>` HTML element.
Styles are automatically scoped within the component, and there is only
one `<button>` HTML element in the component, so the styles can
reference the `<button>` HTML element directly.