Add Astro integration to build site search index
This commit will add an Astro integration to build a site search index
with Pagefind and that will build the search index after the rest of the
build process is done, using the `"astro:build:done"` hook. This is a
similar approach to the one taken by the Astro Starlight docs framework.

https://pagefind.app/
https://starlight.astro.build/guides/site-search/
https://github.com/withastro/starlight/blob/bf939947a3b219a75cd7ac8ade27694c8a927937/packages/starlight/index.ts#L115-L127