Add Astro component for site search UI element
This commit will add an Astro component for site search with Pagefind,
and will add the component to the nav at the top of the page.

The approach taken for the site search component is similar to the one
taken by the Astro Starlight docs framework. A dialog element for search
is shown in the nav at the top of the page. The dialog is a magnifying
glass icon on small viewports, and expands to a text entry dialog with
"Search" and the appropriate keyboard shortcut on larger viewports. The
dialog is implemented as a custom element (web component). The search
depends on JavaScript to function, so a `<template>` HTML element will
be used to set up a document fragment that is only inserted/injected
into the DOM when JavaScript runs. If JavaScript is disabled, the search
dialog will not appear.

Once the search dialog is activated, a larger dialog opens. This dialog
is based on the Pagefind "Default UI" element. Pagefind outputs the code
for the Default UI after building the search index, but this code is not
very useful because it is not available before the site builds. The
auxiliary package `@pagefind/default-ui` must therefore be installed so
the Default UI code is available when checking and building the site.
The Default UI includes a stylesheet `@pagefind/default-ui/css/ui.css`.
Starlight does not use the Default UI stylesheet, but the implementation
included with this commit will use the Default UI stylesheet. Additional
styling is included to match the Default UI with the rest of the site.

The search component `Search.astro` will be incorporated into the nav at
the top of the page. A `<div>` will be added to the nav to group search
and the menu toggle button together and preserve the existing layout.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/template
https://pagefind.app/
https://starlight.astro.build/guides/site-search/
https://github.com/withastro/starlight/blob/3c8b4bb7c8713c4f4820e44fe911969910c67ecd/packages/starlight/components/Search.astro