Adjust gap and padding in nav menu footer
This commit will adjust the gap and padding values used in the nav menu
footer (where the socials and the theme toggle icon are), resulting in
simpler sizing values and improved appearance on intermediate viewport
widths, such as the iPad Pro (1024 px wide).

The global CSS stylesheet contains a reference to a CSS custom property
(variable) named `--icon-padding` but the variable is not defined there.
The variable is actually only defined and used in the nav menu footer.
This commit will remove the `--icon-padding` reference from the global
stylesheet and move it to `Nav.astro` where the variable is defined.