Move nav template `<noscript>` section to bottom
The `Nav.astro` component has a `<noscript>` section that structures the
nav for browsers in which JavaScript is disabled. The `<noscript>` HTML
elements were stuck in the middle of the template, making it difficult
to read. This commit will move the `<noscript>` HTML elements to the
bottom of the template for better readability without changing the
existing functionality.