Improve no-JS display of `Contact.astro`
The `Contact.astro` component includes JavaScript that renders a link
when a user clicks a button. In no-JavaScript browsers, the component
currently has an additional `<noscript>` element explaining that the
link requires JavaScript to function. If the link doesn't function
without JavaScript, it would be preferable to simply not show it at all.

This commit will improve display of `Contact.astro` in no-JS browsers.
A `<template>` HTML element will be used to set up a document fragment
that is only inserted/injected into the DOM when JavaScript runs. When
JavaScript is disabled, there will simply be a bottom border rendered.

https://developer.mozilla.org/en-US/docs/Web/API/DocumentFragment
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/template