Fix fonts in `Search.astro`
In `Search.astro`, a CSS variable `--pagefind-ui-font: var(--font-body)`
is used to control the font within the Pagefind search component, but
there are some other elements in `Search.astro` that are not using the
same font. This commit will update the CSS in `Search.astro` so the font
matches throughout the component.