Format Git commit messages with Prettier
Git commits have a suggested format that can be seen in `.gitmessage`.

- The commit title is limited to 50 characters.
- The commit message body has lines wrapped at 72 characters.

This means that, if re-formatting other prose sources like GitHub PR
descriptions, the user has to manually wrap lines, which is tedious.
VSCode doesn't format commit messages (microsoft/vscode#2718). GitHub
does a brute-force line wrap when reformatting PR descriptions as squash
commit messages, but the format is not particularly readable.

GitHub PR descriptions are formatted as Markdown, so it could make sense
to also format commit messages as Markdown. Prettier is used as the
Markdown formatter in this repo. This commit will configure Prettier as
a Git commit message formatter. To do this, an override is added to the
Prettier configuration file (`.prettierrc`) for `COMMIT_EDITMSG` files
(this is the name of the file Git uses for commit messages). The options
will include `"printWidth": 72` and `"proseWrap": "always"` to conform
to the typical "50/72" Git commit message style. To format with VSCode,
open the Command Palette and select "Format Document (forced)".

https://git-scm.com/docs/git-commit
br3ndonland/dotfiles@6225082